Convert to autobuilding DockerFile.
By using the golang onbuild image it was very easy to create a
DockerFile that could compile the code.  This would allow the image to
be autobuilt by DockerHub.  Unfortunately the image is 760MB in size
compared to just 11MB for the 2 step build process of compiling and
adding only the binary to the docker image.

Since we aren't going to be using DockerHub much longer I don't think that
autobuilding through Dockerhub is worth the extra size.  AWS ECR at this
time does not have an autobuilding feature.  If docker images will be
built with Jenkins jobs the 2 step process should be fine.