Add date to docker tag.
To ensure the docker image contains an recent set of trusted
certificates a new image must be created periodically.  Since this
project is unlikely to be updated very often, it is likely that
consecutive docker images will have the same git SHA.  Prior to this
commit, such images would also have the same docker tag.  This change
adds the date to the docker tag.

An alternative to this would be to have the jenkins job perform a git
commit with the updated certs file, however IMHO this solution seems equally
valid.